requirements.txt
langgraph>=0.2.0
langchain>=0.3.0
langchain-core>=0.1.0
langchain-community>=0.3.0
langchain-text-splitters>=0.0.1
langchain-huggingface>=0.0.1
langchain-mistralai>=0.2.0
langchain-openai>=0.2.0
faiss-cpu>=1.8.0
sentence-transformers>=3.0.0
ddgs>=3.9.0
pypdf>=4.0.0
python-dotenv>=1.0.0
pyspellchecker>=0.8.0
fastapi>=0.115.0
uvicorn>=0.30.0
pydantic>=2.0.0
tqdm>=4.66.0
torch>=2.0.0
httpx>=0.27.0
requests>=2.31.0
beautifulsoup4>=4.12.0
